objects/pickup: merge Pickup_Trigger in TRX
