+++ src/libtrx/game/objects/general/pickup.c
@@ -1,0 +1,14 @@
+#include "game/objects/general/pickup.h"
+
+#include "game/items.h"
+
+bool Pickup_Trigger(const int16_t item_num)
+{
+    ITEM *const item = Item_Get(item_num);
+    if (item->status != IS_INVISIBLE) {
+        return false;
+    }
+
+    item->status = IS_DEACTIVATED;
+    return true;
+}
