+++ src/libtrx/include/libtrx/game/objects/general/pickup.h
@@ -2,5 +2,5 @@
 
 #include "../types.h"
 
-extern bool Pickup_Trigger(int16_t item_num);
+bool Pickup_Trigger(int16_t item_num);
 const OBJECT_BOUNDS *Pickup_Bounds(void);
